last piece of request bin
