+++ core/requestbin/api.go
@@ -1,26 +1,28 @@
-package app
+package requestbin
 
 import (
 	"encoding/json"
+	"fmt"
 	"io/ioutil"
 	"net/http"
+	"strconv"
 	"strings"
 	"time"
 
 	"github.com/Z-M-Huang/Tools/core"
 	"github.com/Z-M-Huang/Tools/data"
-	"github.com/Z-M-Huang/Tools/data/apidata/application"
 	"github.com/Z-M-Huang/Tools/data/db"
-	webData "github.com/Z-M-Huang/Tools/data/webdata/application"
-	applicationlogic "github.com/Z-M-Huang/Tools/logic/application"
 	"github.com/Z-M-Huang/Tools/utils"
 	"github.com/gin-gonic/gin"
 )
 
+//API request bin
+type API struct{}
+
 //CreateRequestBin /api/request-bin/Create
-func CreateRequestBin(c *gin.Context) {
+func (API) CreateRequestBin(c *gin.Context) {
 	response := c.Keys[utils.ResponseCtxKey].(*data.Response)
-	request := &application.CreateBinRequest{}
+	request := &CreateBinRequest{}
 	err := c.ShouldBind(&request)
 	if err != nil {
 		response.SetAlert(&data.AlertData{
@@ -31,13 +33,13 @@ func CreateRequestBin(c *gin.Context) {
 		return
 	}
 
-	bin := applicationlogic.NewRequestBinHistory(request.IsPrivate)
+	bin := NewRequestBinHistory(request.IsPrivate)
 	if bin == nil {
 		core.WriteUnexpectedError(c, response)
 		c.Abort()
 		return
 	}
-	result := &application.CreateBinResponse{
+	result := &CreateBinResponse{
 		ID:              bin.ID,
 		VerificationKey: bin.VerificationKey,
 	}
@@ -47,20 +49,20 @@ func CreateRequestBin(c *gin.Context) {
 }
 
 //RequestIn /api/request-bin/receive/:id
-func RequestIn(c *gin.Context) {
+func (API) RequestIn(c *gin.Context) {
 	id := c.Param("id")
 	if id == "" {
 		c.AbortWithStatus(http.StatusNotFound)
 		return
 	}
 
-	binData := applicationlogic.GetRequestBinHistory(id)
+	binData := GetRequestBinHistory(id)
 	if binData == nil {
 		c.AbortWithStatus(http.StatusNotFound)
 		return
 	}
 
-	history := &webData.RequestHistory{
+	history := &History{
 		TimeReceived:        time.Now().UTC(),
 		Method:              c.Request.Method,
 		Proto:               c.Request.Proto,
@@ -112,11 +114,11 @@ func RequestIn(c *gin.Context) {
 	}
 	history.Body = string(bodyBytes)
 
-	binData.History = append([]*webData.RequestHistory{history}, binData.History...)
+	binData.History = append([]*History{history}, binData.History...)
 	if len(binData.History) >= 20 {
 		binData.History = binData.History[0:19]
 	}
-	key := applicationlogic.GetRequestBinKey(binData.ID)
+	key := GetRequestBinKey(binData.ID)
 	err = db.RedisSetBytes(key, binData, 24*time.Hour)
 	if err != nil {
 		utils.Logger.Error(err.Error())
@@ -124,4 +126,52 @@ func RequestIn(c *gin.Context) {
 		return
 	}
 	c.String(200, "Okay")
+}
+
+//GetRequestBinHistory get request history by id
+func GetRequestBinHistory(id string) *PageData {
+	key := GetRequestBinKey(id)
+	data := &PageData{}
+	err := db.RedisGet(key, data)
+	if err != nil {
+		utils.Logger.Error(err.Error())
+		return nil
+	}
+	return data
+}
+
+//NewRequestBinHistory new request history
+func NewRequestBinHistory(private bool) *PageData {
+	binData := &PageData{
+		ID: strconv.FormatInt(time.Now().Unix(), 10),
+	}
+	if data.Config.HTTPS {
+		binData.URL = "https://"
+	} else {
+		binData.URL = "http://"
+	}
+
+	binData.URL += data.Config.Host + "/api/request-bin/receive/" + binData.ID
+
+	if private {
+		binData.VerificationKey = utils.RandomString(30)
+	}
+
+	bytes, err := json.Marshal(binData)
+	if err != nil {
+		utils.Logger.Error(err.Error())
+		return nil
+	}
+	key := GetRequestBinKey(binData.ID)
+	err = db.RedisSet(key, bytes, 24*time.Hour)
+	if err != nil {
+		utils.Logger.Error(err.Error())
+		return nil
+	}
+	return binData
+}
+
+//GetRequestBinKey request bin key
+func GetRequestBinKey(id string) string {
+	return fmt.Sprintf("APP_REQUEST_BIN_%s", id)
 }
