+++ data/apidata/application/request_bin.go
@@ -1,12 +1,1 @@
 package application
-
-//CreateBinRequest /api/request-bin/create
-type CreateBinRequest struct {
-	IsPrivate bool `json:"isPrivate" xml:"isPrivate" form:"isPrivate"`
-}
-
-//CreateBinResponse response
-type CreateBinResponse struct {
-	ID              string
-	VerificationKey string
-}
